requirements.txt
Flask
Pillow
numpy
svgwrite
scikit-image
